requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
paho-mqtt==2.1.0
python-osc==1.8.3
Werkzeug==3.1.3
zipp==3.21.0
requests==2.32.3
google-generativeai==0.8.4
python-dotenv>=1.0.0
